requirements.txt
nextcord
ffmpeg-python
python-dotenv
